clean Monitoring.h
